+++ driver-core/src/main/java/com/datastax/driver/core/Cluster.java
@@ -29,6 +29,7 @@ import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.MapMaker;
 import com.google.common.collect.SetMultimap;
+import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.*;
 
 import org.slf4j.Logger;
@@ -61,6 +62,10 @@ import com.datastax.driver.core.policies.*;
 public class Cluster implements Closeable {
 
     private static final Logger logger = LoggerFactory.getLogger(Cluster.class);
+
+    private static final int NEW_NODE_DELAY_SECONDS = SystemProperties.getInt("com.datastax.driver.NEW_NODE_DELAY_SECONDS", 1);
+    private static final int NON_BLOCKING_EXECUTOR_SIZE = SystemProperties.getInt("com.datastax.driver.NON_BLOCKING_EXECUTOR_SIZE",
+                                                                                  Runtime.getRuntime().availableProcessors());
 
     // Some per-JVM number that allows to generate unique cluster names when
     // multiple Cluster instance are created in the same JVM.
@@ -1093,6 +1098,8 @@ public class Cluster implements Closeable {
         // An executor for tasks that might block some time, like creating new connection, but are generally not too critical.
         final ListeningExecutorService blockingExecutor;
 
+        final ConnectionReaper reaper;
+
         final AtomicReference<CloseFuture> closeFuture = new AtomicReference<CloseFuture>();
 
         // All the queries that have been prepared (we keep them so we can re-prepared them when a node fail or a
@@ -1111,8 +1118,10 @@ public class Cluster implements Closeable {
             this.configuration = configuration;
             this.configuration.register(this);
 
-            this.executor = makeExecutor(Runtime.getRuntime().availableProcessors(), "Cassandra Java Driver worker-%d");
+            this.executor = makeExecutor(NON_BLOCKING_EXECUTOR_SIZE, "Cassandra Java Driver worker-%d");
             this.blockingExecutor = makeExecutor(2, "Cassandra Java Driver blocking tasks worker-%d");
+
+            this.reaper = new ConnectionReaper();
 
             this.metadata = new Metadata(this);
             this.contactPoints = contactPoints;
@@ -1190,7 +1199,8 @@ public class Cluster implements Closeable {
 
         InetSocketAddress translateAddress(InetAddress address) {
             InetSocketAddress sa = new InetSocketAddress(address, connectionFactory.getPort());
-            return configuration.getPolicies().getAddressTranslater().translate(sa);
+            InetSocketAddress translated = configuration.getPolicies().getAddressTranslater().translate(sa);
+            return translated == null ? sa : translated;
         }
 
         private Session newSession() {
@@ -1253,6 +1263,12 @@ public class Cluster implements Closeable {
                       + "Cluster.Builder#usingProtocolVersion() when creating the Cluster instance.", host);
         }
 
+        void logClusterNameMismatch(Host host, String expectedClusterName, String actualClusterName) {
+            logger.warn("Detected added or restarted Cassandra host {} but ignoring it since its cluster name '{}' does not match the one "
+                        + "currently known ({})",
+                        host, actualClusterName, expectedClusterName);
+        }
+
         public ListenableFuture<?> triggerOnUp(final Host host) {
             return executor.submit(new ExceptionCatchingRunnable() {
                 @Override
@@ -1300,6 +1316,9 @@ public class Cluster implements Closeable {
                 // Don't propagate because we don't want to prevent other listener to run
             } catch (UnsupportedProtocolVersionException e) {
                 logUnsupportedVersionProtocol(host);
+                return;
+            } catch (ClusterNameMismatchException e) {
+                logClusterNameMismatch(host, e.expectedClusterName, e.actualClusterName);
                 return;
             }
 
@@ -1473,7 +1492,7 @@ public class Cluster implements Closeable {
             logger.debug("{} is down, scheduling connection retries", host);
             new AbstractReconnectionHandler(reconnectionExecutor, reconnectionPolicy().newSchedule(), host.reconnectionAttempt) {
 
-                protected Connection tryReconnect() throws ConnectionException, InterruptedException, UnsupportedProtocolVersionException {
+                protected Connection tryReconnect() throws ConnectionException, InterruptedException, UnsupportedProtocolVersionException, ClusterNameMismatchException {
                     return connectionFactory.open(host);
                 }
 
@@ -1482,19 +1501,22 @@ public class Cluster implements Closeable {
                     // TODO: this is a bit wasteful, we should consider passing it to onAdd/onUp so
                     // we use it for the first HostConnectionPool created
                     connection.closeAsync();
-                    logger.debug("Successful reconnection to {}, setting host UP", host);
                     // Make sure we have up-to-date infos on that host before adding it (so we typically
                     // catch that an upgraded node uses a new cassandra version).
-                    controlConnection.refreshNodeInfo(host);
-                    try {
-                        if (isHostAddition)
-                            onAdd(host);
-                        else
-                            onUp(host);
-                    } catch (InterruptedException e) {
-                        Thread.currentThread().interrupt();
-                    } catch (Exception e) {
-                        logger.error("Unexpected error while setting node up", e);
+                    if (controlConnection.refreshNodeInfo(host)) {
+                        logger.debug("Successful reconnection to {}, setting host UP", host);
+                        try {
+                            if (isHostAddition)
+                                onAdd(host);
+                            else
+                                onUp(host);
+                        } catch (InterruptedException e) {
+                            Thread.currentThread().interrupt();
+                        } catch (Exception e) {
+                            logger.error("Unexpected error while setting node up", e);
+                        }
+                    } else {
+                        logger.debug("Not enough info for {}, ignoring host", host);
                     }
                 }
 
@@ -1505,7 +1527,7 @@ public class Cluster implements Closeable {
                 }
 
                 protected boolean onUnknownException(Exception e, long nextDelayMs) {
-                    logger.error(String.format("Unknown error during control connection reconnection, scheduling retry in %d milliseconds", nextDelayMs), e);
+                    logger.error(String.format("Unknown error during reconnection to %s, scheduling retry in %d milliseconds", host, nextDelayMs), e);
                     return true;
                 }
 
@@ -1557,6 +1579,9 @@ public class Cluster implements Closeable {
                 // Don't propagate because we don't want to prevent other listener to run
             } catch (UnsupportedProtocolVersionException e) {
                 logUnsupportedVersionProtocol(host);
+                return;
+            } catch (ClusterNameMismatchException e) {
+                logClusterNameMismatch(host, e.expectedClusterName, e.actualClusterName);
                 return;
             }
 
@@ -1685,7 +1710,7 @@ public class Cluster implements Closeable {
             return stmt;
         }
 
-        private void prepareAllQueries(Host host) throws InterruptedException, UnsupportedProtocolVersionException {
+        private void prepareAllQueries(Host host) throws InterruptedException, UnsupportedProtocolVersionException, ClusterNameMismatchException {
             if (preparedQueries.isEmpty())
                 return;
 
@@ -1773,7 +1798,7 @@ public class Cluster implements Closeable {
                         // Before refreshing the schema, wait for schema agreement so
                         // that querying a table just after having created it don't fail.
                         if (!ControlConnection.waitForSchemaAgreement(connection, Cluster.Manager.this))
-                            logger.warn("No schema agreement from live replicas after {} ms. The schema may not be up to date on some nodes.", ControlConnection.MAX_SCHEMA_AGREEMENT_WAIT_MS);
+                            logger.warn("No schema agreement from live replicas after {} s. The schema may not be up to date on some nodes.", ControlConnection.MAX_SCHEMA_AGREEMENT_WAIT_SECONDS);
                         ControlConnection.refreshSchema(connection, keyspace, table, Cluster.Manager.this, false);
                     } catch (Exception e) {
                         logger.error("Error during schema refresh ({}). The schema from Cluster.getMetadata() might appear stale. Asynchronously submitting job to fix.", e.getMessage());
@@ -1809,18 +1834,21 @@ public class Cluster implements Closeable {
                             final Host newHost = metadata.add(tpAddr);
                             if (newHost != null) {
                                 // Cassandra tends to send notifications for new/up nodes a bit early (it is triggered once
-                                // gossip is up, but that is before the client-side server is up), so we add a somewhat random
-                                // 1 second delay (otherwise the connection will likely fail and have to be retry which is
-                                // wasteful). This probably should be fixed C* side, after which we'll be able to remove this.
+                                // gossip is up, but that is before the client-side server is up), so we add a delay
+                                // (otherwise the connection will likely fail and have to be retry which is wasteful). This
+                                // probably should be fixed C* side, after which we'll be able to remove this.
                                 scheduledTasksExecutor.schedule(new ExceptionCatchingRunnable() {
                                     @Override
                                     public void runMayThrow() throws InterruptedException, ExecutionException {
                                         // Make sure we have up-to-date infos on that host before adding it (so we typically
                                         // catch that an upgraded node uses a new cassandra version).
-                                        controlConnection.refreshNodeInfo(newHost);
-                                        onAdd(newHost);
+                                        if (controlConnection.refreshNodeInfo(newHost)) {
+                                            onAdd(newHost);
+                                        } else {
+                                            logger.debug("Not enough info for {}, ignoring host", newHost);
+                                        }
                                     }
-                                }, 1, TimeUnit.SECONDS);
+                                }, NEW_NODE_DELAY_SECONDS, TimeUnit.SECONDS);
                             }
                             break;
                         case REMOVED_NODE:
@@ -1856,18 +1884,24 @@ public class Cluster implements Closeable {
                                     public void runMayThrow() throws InterruptedException, ExecutionException {
                                         // Make sure we have up-to-date infos on that host before adding it (so we typically
                                         // catch that an upgraded node uses a new cassandra version).
-                                        controlConnection.refreshNodeInfo(h);
-                                        onAdd(h);
+                                        if (controlConnection.refreshNodeInfo(h)) {
+                                            onAdd(h);
+                                        } else {
+                                            logger.debug("Not enough info for {}, ignoring host", h);
+                                        }
                                     }
-                                }, 1, TimeUnit.SECONDS);
+                                }, NEW_NODE_DELAY_SECONDS, TimeUnit.SECONDS);
                             } else {
                                 executor.submit(new ExceptionCatchingRunnable() {
                                     @Override
                                     public void runMayThrow() throws InterruptedException, ExecutionException {
                                         // Make sure we have up-to-date infos on that host before adding it (so we typically
                                         // catch that an upgraded node uses a new cassandra version).
-                                        controlConnection.refreshNodeInfo(hostUp);
-                                        onUp(hostUp);
+                                        if (controlConnection.refreshNodeInfo(hostUp)) {
+                                            onUp(hostUp);
+                                        } else {
+                                            logger.debug("Not enough info for {}, ignoring host", hostUp);
+                                        }
                                     }
                                 });
                             }
@@ -1957,6 +1991,8 @@ public class Cluster implements Closeable {
                             // connectionFactory at the very last
                             connectionFactory.shutdown();
 
+                            reaper.shutdown();
+
                             set(null);
                         } catch (InterruptedException e) {
                             Thread.currentThread().interrupt();
@@ -1965,6 +2001,63 @@ public class Cluster implements Closeable {
                     }
                 }).start();
             }
+        }
+    }
+
+    /**
+     * Periodically ensures that closed connections are properly terminated once they have no more pending requests.
+     *
+     * This is normally done when the connection errors out, or when the last request is processed; this class acts as
+     * a last-effort protection since unterminated connections can lead to deadlocks. If it terminates a connection,
+     * this indicates a bug; warnings are logged so that this can be reported.
+     */
+    static class ConnectionReaper {
+        private static final int INTERVAL_MS = 15000;
+
+        private final ScheduledExecutorService executor = Executors.newScheduledThreadPool(1, threadFactory("Reaper-%d"));
+        private final Set<Connection> connections = Sets.newConcurrentHashSet();
+
+        private volatile boolean shutdown;
+
+        private final Runnable reaperTask = new Runnable() {
+            @Override
+            public void run() {
+                reapConnections();
+                if (!executor.isShutdown())
+                    executor.schedule(this, INTERVAL_MS, TimeUnit.MILLISECONDS);
+            }
+
+            private final void reapConnections() {
+                Iterator<Connection> iterator = connections.iterator();
+                while (iterator.hasNext()) {
+                    Connection connection = iterator.next();
+                    boolean terminated = connection.terminate(false, true);
+                    if (terminated)
+                        iterator.remove();
+                }
+            }
+        };
+
+        ConnectionReaper() {
+            executor.schedule(reaperTask, INTERVAL_MS, TimeUnit.MILLISECONDS);
+        }
+
+        void register(Connection connection) {
+            if (shutdown) {
+                // This should not happen since the reaper is shut down after all sessions.
+                logger.warn("Connection registered after reaper shutdown: {}", connection);
+                connection.terminate(true, true);
+            } else {
+                connections.add(connection);
+            }
+        }
+
+        void shutdown() {
+            shutdown = true;
+            executor.shutdown();
+            // Run the task one last time if a call to register raced with the two lines above (i.e.
+            // saw shutdown as false but added its connection after the last task in the executor ran).
+            reaperTask.run();
         }
     }
 }