+++ driver-core/src/main/java/com/datastax/driver/core/PooledConnection.java
@@ -16,6 +16,7 @@
 package com.datastax.driver.core;
 
 import java.net.InetSocketAddress;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * A connection that is associated to a pool.
@@ -24,7 +25,10 @@ class PooledConnection extends Connection {
 
     private final HostConnectionPool pool;
 
-    PooledConnection(String name, InetSocketAddress address, Factory factory, HostConnectionPool pool) throws ConnectionException, InterruptedException, UnsupportedProtocolVersionException {
+    /** Used in {@link HostConnectionPool} to handle races between two threads trying to trash the same connection */
+    final AtomicBoolean markForTrash = new AtomicBoolean();
+
+    PooledConnection(String name, InetSocketAddress address, Factory factory, HostConnectionPool pool) throws ConnectionException, InterruptedException, UnsupportedProtocolVersionException, ClusterNameMismatchException {
         super(name, address, factory);
         this.pool = pool;
     }
@@ -47,7 +51,7 @@ class PooledConnection extends Connection {
         if (hostIsDown) {
             pool.closeAsync().force();
         } else {
-            pool.replace(this);
+            pool.replaceDefunctConnection(this);
         }
     }
 }