+++ driver-core/src/test/java/com/datastax/driver/core/TypeCodecTest.java
@@ -17,6 +17,11 @@ package com.datastax.driver.core;
 
 import java.nio.ByteBuffer;
 import java.util.Map;
+import java.util.Collections;
+import java.util.List;
+
+import com.google.common.collect.Lists;
+import com.google.common.base.Strings;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -49,5 +54,21 @@ public class TypeCodecTest {
     public void testCustomValueMap() throws Exception {
         TypeCodec<Map<String, ByteBuffer>> mapType = TypeCodec.mapOf(text(), CUSTOM_FOO, 2);
         Assert.assertNotNull(mapType);
+    }
+
+    @Test(groups = "unit", expectedExceptions = { IllegalArgumentException.class })
+    public void collectionTooLargeTest() throws Exception {
+        TypeCodec<List<Integer>> listType = TypeCodec.listOf(DataType.cint(), 2);
+        List<Integer> list = Collections.nCopies(65536, 1);
+
+        listType.serialize(list);
+    }
+
+    @Test(groups = "unit", expectedExceptions = { IllegalArgumentException.class })
+    public void collectionElementTooLargeTest() throws Exception {
+        TypeCodec<List<String>> listType = TypeCodec.listOf(DataType.text(), 2);
+        List<String> list = Lists.newArrayList(Strings.repeat("a", 65536));
+
+        listType.serialize(list);
     }
 }