+++ driver-core/CHANGELOG.rst
@@ -1,6 +1,20 @@
 CHANGELOG
 =========
 
+2.1.1:
+------
+
+Merged from 2.0 branch:
+
+- [bug] Check cluster name when connecting to a new node (JAVA-397)
+- [bug] Add missing CAS delete support in QueryBuilder (JAVA-326)
+- [bug] Add collection and data length checks during serialization (JAVA-363)
+- [improvement] Surface number of retries in metrics (JAVA-329)
+- [bug] Do not use a host when no rpc_address found for it (JAVA-428)
+- [improvement] Add ResultSet.wasApplied() for conditional queries (JAVA-358)
+- [bug] Fix negative HostConnectionPool open count (JAVA-349)
+
+
 2.1.0:
 ------
 
@@ -44,6 +58,29 @@ Merged from 2.0 branch: everything up to 2.0.3 (included), and the following.
 - [bug] Make metadata parsing more lenient (JAVA-377, JAVA-391)
 
 
+2.0.6:
+------
+
+- [bug] Check cluster name when connecting to a new node (JAVA-397)
+- [bug] Add missing CAS delete support in QueryBuilder (JAVA-326)
+- [bug] Add collection and data length checks during serialization (JAVA-363)
+- [improvement] Surface number of retries in metrics (JAVA-329)
+- [bug] Do not use a host when no rpc_address found for it (JAVA-428)
+- [improvement] Add ResultSet.wasApplied() for conditional queries (JAVA-358)
+- [bug] Fix negative HostConnectionPool open count (JAVA-349)
+
+
+2.0.5:
+------
+
+- [bug] Release connections on ResultSetFuture#cancel (JAVA-407)
+- [bug] Fix handling of SimpleStatement with values in query builder
+  batches (JAVA-393)
+- [bug] Ensure pool is properly closed in onDown (JAVA-417)
+- [bug] Fix tokenMap initialization at startup (JAVA-415)
+- [bug] Avoid deadlock on close (JAVA-418)
+
+
 2.0.4:
 ------
 
@@ -74,7 +111,7 @@ Merged from 2.0 branch: everything up to 2.0.3 (included), and the following.
 - [bug] Signal connection failure sooner to avoid missing them
   (JAVA-367)
 - [bug] Throw UnsupportedOperationException for protocol batch
-  setSerialCL (JAVA-377)
+  setSerialCL (JAVA-337)
 
 Merged from 1.0 branch: